harden and cleanup secondary index
