--- doradb-storage/src/index/composite_secondary_index.rs
@@ -1,1241 +1,0 @@
-//! Composite MemTree/DiskTree secondary-index core.
-//!
-//! The types in this module are the RFC 0014 phase-3 user-table core only.
-//! They group the current in-memory BTree-backed secondary index with the
-//! current published secondary DiskTree root and preserve the existing unique
-//! and non-unique trait contracts without wiring the composite into table
-//! runtime.
-
-use super::disk_tree::{
-    NonUniqueDiskTree, NonUniqueDiskTreeRuntime, UniqueDiskTree, UniqueDiskTreeRuntime,
-};
-use super::non_unique_index::{GenericNonUniqueBTreeIndex, NonUniqueIndex, NonUniqueMemTreeEntry};
-use super::secondary_index::{GenericSecondaryIndex, IndexCompareExchange, IndexInsert};
-use super::unique_index::{GenericUniqueBTreeIndex, UniqueIndex, UniqueMemTreeEntry};
-use crate::buffer::{BufferPool, PoolGuard, ReadonlyBufferPool};
-use crate::catalog::TableMetadata;
-use crate::error::{Error, Result};
-use crate::file::cow_file::BlockID;
-use crate::file::table_file::TableFile;
-use crate::index::util::Maskable;
-use crate::quiescent::QuiescentGuard;
-use crate::row::RowID;
-use crate::trx::TrxID;
-use crate::value::Val;
-use std::cmp::Ordering;
-use std::sync::Arc;
-
-/// Runtime cold-layer opener for one user-table secondary DiskTree.
-///
-/// The runtime is table-specific by construction. Each open reads the currently
-/// published secondary root from the table file, then hands that copied root to
-/// a typed DiskTree reader. Already opened readers keep their root snapshot.
-pub(crate) struct SecondaryDiskTreeRuntime {
-    index_no: usize,
-    table_file: Arc<TableFile>,
-    kind: SecondaryDiskTreeRuntimeKind,
-}
-
-enum SecondaryDiskTreeRuntimeKind {
-    Unique(UniqueDiskTreeRuntime),
-    NonUnique(NonUniqueDiskTreeRuntime),
-}
-
-impl SecondaryDiskTreeRuntime {
-    /// Create a cold-layer runtime for one table secondary index.
-    #[inline]
-    pub(crate) fn new(
-        index_no: usize,
-        metadata: Arc<TableMetadata>,
-        table_file: Arc<TableFile>,
-        disk_pool: QuiescentGuard<ReadonlyBufferPool>,
-    ) -> Result<Self> {
-        let index_spec = metadata
-            .index_specs
-            .get(index_no)
-            .ok_or(Error::InvalidArgument)?;
-        table_file
-            .active_root()
-            .secondary_index_roots
-            .get(index_no)
-            .ok_or(Error::InvalidArgument)?;
-        let file_kind = table_file.file_kind();
-        let file = Arc::clone(table_file.sparse_file());
-        let kind = if index_spec.unique() {
-            SecondaryDiskTreeRuntimeKind::Unique(UniqueDiskTreeRuntime::new(
-                index_spec,
-                metadata.as_ref(),
-                file_kind,
-                file,
-                disk_pool,
-            )?)
-        } else {
-            SecondaryDiskTreeRuntimeKind::NonUnique(NonUniqueDiskTreeRuntime::new(
-                index_spec,
-                metadata.as_ref(),
-                file_kind,
-                file,
-                disk_pool,
-            )?)
-        };
-        let runtime = Self {
-            index_no,
-            table_file,
-            kind,
-        };
-        Ok(runtime)
-    }
-
-    /// Return the current published DiskTree root for this secondary index.
-    #[inline]
-    pub(crate) fn published_root(&self) -> Result<BlockID> {
-        self.table_file
-            .active_root()
-            .secondary_index_roots
-            .get(self.index_no)
-            .copied()
-            .ok_or(Error::InvalidArgument)
-    }
-
-    /// Return the table index number represented by this context.
-    #[inline]
-    pub(crate) fn index_no(&self) -> usize {
-        self.index_no
-    }
-
-    #[inline]
-    pub(crate) fn is_unique(&self) -> bool {
-        matches!(self.kind, SecondaryDiskTreeRuntimeKind::Unique(_))
-    }
-
-    #[inline]
-    pub(crate) fn disk_pool_guard(&self) -> PoolGuard {
-        match &self.kind {
-            SecondaryDiskTreeRuntimeKind::Unique(runtime) => runtime.disk_pool_guard(),
-            SecondaryDiskTreeRuntimeKind::NonUnique(runtime) => runtime.disk_pool_guard(),
-        }
-    }
-
-    #[inline]
-    pub(crate) fn open_unique<'a>(
-        &'a self,
-        disk_pool_guard: &'a PoolGuard,
-    ) -> Result<UniqueDiskTree<'a>> {
-        self.open_unique_at(self.published_root()?, disk_pool_guard)
-    }
-
-    #[inline]
-    pub(crate) fn open_unique_at<'a>(
-        &'a self,
-        root_block_id: BlockID,
-        disk_pool_guard: &'a PoolGuard,
-    ) -> Result<UniqueDiskTree<'a>> {
-        match &self.kind {
-            SecondaryDiskTreeRuntimeKind::Unique(runtime) => {
-                Ok(runtime.open(root_block_id, disk_pool_guard))
-            }
-            SecondaryDiskTreeRuntimeKind::NonUnique(_) => Err(Error::InvalidArgument),
-        }
-    }
-
-    #[inline]
-    pub(crate) fn open_non_unique<'a>(
-        &'a self,
-        disk_pool_guard: &'a PoolGuard,
-    ) -> Result<NonUniqueDiskTree<'a>> {
-        self.open_non_unique_at(self.published_root()?, disk_pool_guard)
-    }
-
-    #[inline]
-    pub(crate) fn open_non_unique_at<'a>(
-        &'a self,
-        root_block_id: BlockID,
-        disk_pool_guard: &'a PoolGuard,
-    ) -> Result<NonUniqueDiskTree<'a>> {
-        match &self.kind {
-            SecondaryDiskTreeRuntimeKind::Unique(_) => Err(Error::InvalidArgument),
-            SecondaryDiskTreeRuntimeKind::NonUnique(runtime) => {
-                Ok(runtime.open(root_block_id, disk_pool_guard))
-            }
-        }
-    }
-}
-
-/// Composite unique secondary index over a mutable MemTree and cold DiskTree.
-pub(crate) struct DualTreeUniqueIndex<P: 'static> {
-    mem: GenericUniqueBTreeIndex<P>,
-    disk: SecondaryDiskTreeRuntime,
-}
-
-impl<P: BufferPool> DualTreeUniqueIndex<P> {
-    /// Return the cold DiskTree runtime used by this composite.
-    #[inline]
-    pub(crate) fn disk(&self) -> &SecondaryDiskTreeRuntime {
-        &self.disk
-    }
-
-    /// Destroy the mutable MemTree owned by this composite index.
-    #[inline]
-    pub(crate) async fn destroy(self, pool_guard: &PoolGuard) -> Result<()> {
-        self.mem.destroy(pool_guard).await
-    }
-
-    /// Insert into the MemTree backend during recovery without probing DiskTree.
-    ///
-    /// Recovery uses this for hot/post-checkpoint row-page state. Checkpointed
-    /// cold state is already represented by the DiskTree root and must not be
-    /// treated as a duplicate while rebuilding MemTree.
-    #[inline]
-    pub(crate) async fn insert_recovery_mem_only(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        row_id: RowID,
-        merge_if_match_deleted: bool,
-        ts: TrxID,
-    ) -> Result<IndexInsert> {
-        self.mem
-            .insert_if_not_exists(pool_guard, key, row_id, merge_if_match_deleted, ts)
-            .await
-    }
-
-    /// Compare-exchange the MemTree backend during recovery without DiskTree access.
-    #[inline]
-    pub(crate) async fn compare_exchange_recovery_mem_only(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        old_row_id: RowID,
-        new_row_id: RowID,
-        ts: TrxID,
-    ) -> Result<IndexCompareExchange> {
-        self.mem
-            .compare_exchange(pool_guard, key, old_row_id, new_row_id, ts)
-            .await
-    }
-}
-
-impl<P: BufferPool + 'static> UniqueIndex for DualTreeUniqueIndex<P> {
-    #[inline]
-    async fn lookup(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        ts: TrxID,
-    ) -> Result<Option<(RowID, bool)>> {
-        if let Some(hit) = self.mem.lookup(pool_guard, key, ts).await? {
-            return Ok(Some(hit));
-        }
-        let disk_pool_guard = self.disk.disk_pool_guard();
-        let disk = self.disk.open_unique(&disk_pool_guard)?;
-        Ok(disk.lookup(key).await?.map(|row_id| (row_id, false)))
-    }
-
-    #[inline]
-    async fn insert_if_not_exists(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        row_id: RowID,
-        merge_if_match_deleted: bool,
-        ts: TrxID,
-    ) -> Result<IndexInsert> {
-        debug_assert!(!row_id.is_deleted());
-        if let Some((old_row_id, deleted)) = self.mem.lookup(pool_guard, key, ts).await? {
-            if merge_if_match_deleted && deleted && old_row_id == row_id {
-                return self
-                    .mem
-                    .insert_if_not_exists(pool_guard, key, row_id, true, ts)
-                    .await;
-            }
-            return Ok(IndexInsert::DuplicateKey(old_row_id, deleted));
-        }
-        let disk_pool_guard = self.disk.disk_pool_guard();
-        let disk = self.disk.open_unique(&disk_pool_guard)?;
-        if let Some(cold_row_id) = disk.lookup(key).await? {
-            return Ok(IndexInsert::DuplicateKey(cold_row_id, false));
-        }
-        self.mem
-            .insert_if_not_exists(pool_guard, key, row_id, false, ts)
-            .await
-    }
-
-    #[inline]
-    async fn compare_delete(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        old_row_id: RowID,
-        ignore_del_mask: bool,
-        ts: TrxID,
-    ) -> Result<bool> {
-        debug_assert!(!old_row_id.is_deleted());
-        if self.mem.lookup(pool_guard, key, ts).await?.is_some() {
-            return self
-                .mem
-                .compare_delete(pool_guard, key, old_row_id, ignore_del_mask, ts)
-                .await;
-        }
-        Ok(true)
-    }
-
-    #[inline]
-    async fn compare_exchange(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        old_row_id: RowID,
-        new_row_id: RowID,
-        ts: TrxID,
-    ) -> Result<IndexCompareExchange> {
-        if self.mem.lookup(pool_guard, key, ts).await?.is_some() {
-            return self
-                .mem
-                .compare_exchange(pool_guard, key, old_row_id, new_row_id, ts)
-                .await;
-        }
-        // Delete-shadows are MemTree overlays. If purge removed the shadow
-        // between the caller's lookup and this compare-exchange, retry instead
-        // of comparing the masked RowID against the unmasked DiskTree owner.
-        if old_row_id.is_deleted() {
-            return Ok(IndexCompareExchange::NotExists);
-        }
-        let disk_pool_guard = self.disk.disk_pool_guard();
-        let disk = self.disk.open_unique(&disk_pool_guard)?;
-        match disk.lookup(key).await? {
-            Some(cold_row_id) if cold_row_id == old_row_id => {
-                if self
-                    .mem
-                    .insert_overlay_if_absent(pool_guard, key, new_row_id, ts)
-                    .await?
-                {
-                    Ok(IndexCompareExchange::Ok)
-                } else {
-                    Ok(IndexCompareExchange::Mismatch)
-                }
-            }
-            Some(_) => Ok(IndexCompareExchange::Mismatch),
-            None => Ok(IndexCompareExchange::NotExists),
-        }
-    }
-
-    #[inline]
-    async fn scan_values(
-        &self,
-        pool_guard: &PoolGuard,
-        values: &mut Vec<RowID>,
-        _ts: TrxID,
-    ) -> Result<()> {
-        let mem_entries = self.mem.scan_encoded_entries(pool_guard).await?;
-        let disk_pool_guard = self.disk.disk_pool_guard();
-        let disk = self.disk.open_unique(&disk_pool_guard)?;
-        let disk_entries = disk.scan_entries().await?;
-        merge_unique_entries(&mem_entries, &disk_entries, values);
-        Ok(())
-    }
-}
-
-/// Composite non-unique secondary index over a mutable MemTree and cold DiskTree.
-pub(crate) struct DualTreeNonUniqueIndex<P: 'static> {
-    mem: GenericNonUniqueBTreeIndex<P>,
-    disk: SecondaryDiskTreeRuntime,
-}
-
-impl<P: BufferPool> DualTreeNonUniqueIndex<P> {
-    /// Return the cold DiskTree runtime used by this composite.
-    #[inline]
-    pub(crate) fn disk(&self) -> &SecondaryDiskTreeRuntime {
-        &self.disk
-    }
-
-    /// Destroy the mutable MemTree owned by this composite index.
-    #[inline]
-    pub(crate) async fn destroy(self, pool_guard: &PoolGuard) -> Result<()> {
-        self.mem.destroy(pool_guard).await
-    }
-
-    /// Insert into the MemTree backend during recovery without probing DiskTree.
-    ///
-    /// Non-unique DiskTree exact entries are checkpointed cold state. Hot redo
-    /// rebuilds only exact MemTree entries and must not reject a row because a
-    /// checkpointed cold exact entry exists.
-    #[inline]
-    pub(crate) async fn insert_recovery_mem_only(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        row_id: RowID,
-        merge_if_match_deleted: bool,
-        ts: TrxID,
-    ) -> Result<IndexInsert> {
-        self.mem
-            .insert_if_not_exists(pool_guard, key, row_id, merge_if_match_deleted, ts)
-            .await
-    }
-}
-
-impl<P: BufferPool + 'static> NonUniqueIndex for DualTreeNonUniqueIndex<P> {
-    #[inline]
-    async fn lookup(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        res: &mut Vec<RowID>,
-        _ts: TrxID,
-    ) -> Result<()> {
-        let mem_entries = self.mem.lookup_encoded_entries(pool_guard, key).await?;
-        let disk_pool_guard = self.disk.disk_pool_guard();
-        let disk = self.disk.open_non_unique(&disk_pool_guard)?;
-        let disk_entries = disk.prefix_scan_entries(key).await?;
-        merge_non_unique_entries(&mem_entries, &disk_entries, res);
-        Ok(())
-    }
-
-    #[inline]
-    async fn lookup_unique(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        row_id: RowID,
-        ts: TrxID,
-    ) -> Result<Option<bool>> {
-        if let Some(mem_hit) = self.mem.lookup_unique(pool_guard, key, row_id, ts).await? {
-            return Ok(Some(mem_hit));
-        }
-        let disk_pool_guard = self.disk.disk_pool_guard();
-        let disk = self.disk.open_non_unique(&disk_pool_guard)?;
-        if disk.contains_exact(key, row_id).await? {
-            Ok(Some(true))
-        } else {
-            Ok(None)
-        }
-    }
-
-    #[inline]
-    async fn insert_if_not_exists(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        row_id: RowID,
-        merge_if_match_deleted: bool,
-        ts: TrxID,
-    ) -> Result<IndexInsert> {
-        debug_assert!(!row_id.is_deleted());
-        if let Some(active) = self.mem.lookup_unique(pool_guard, key, row_id, ts).await? {
-            if merge_if_match_deleted && !active {
-                return self
-                    .mem
-                    .insert_if_not_exists(pool_guard, key, row_id, true, ts)
-                    .await;
-            }
-            return Ok(IndexInsert::DuplicateKey(row_id, !active));
-        }
-        let disk_pool_guard = self.disk.disk_pool_guard();
-        let disk = self.disk.open_non_unique(&disk_pool_guard)?;
-        if disk.contains_exact(key, row_id).await? {
-            return Ok(IndexInsert::DuplicateKey(row_id, false));
-        }
-        self.mem
-            .insert_if_not_exists(pool_guard, key, row_id, false, ts)
-            .await
-    }
-
-    #[inline]
-    async fn mask_as_deleted(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        row_id: RowID,
-        ts: TrxID,
-    ) -> Result<bool> {
-        debug_assert!(!row_id.is_deleted());
-        match self.mem.lookup_unique(pool_guard, key, row_id, ts).await? {
-            Some(true) => self.mem.mask_as_deleted(pool_guard, key, row_id, ts).await,
-            Some(false) => Ok(false),
-            None => {
-                let disk_pool_guard = self.disk.disk_pool_guard();
-                let disk = self.disk.open_non_unique(&disk_pool_guard)?;
-                if disk.contains_exact(key, row_id).await? {
-                    self.mem
-                        .insert_delete_overlay_if_absent(pool_guard, key, row_id, ts)
-                        .await
-                } else {
-                    Ok(false)
-                }
-            }
-        }
-    }
-
-    #[inline]
-    async fn mask_as_active(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        row_id: RowID,
-        ts: TrxID,
-    ) -> Result<bool> {
-        self.mem.mask_as_active(pool_guard, key, row_id, ts).await
-    }
-
-    #[inline]
-    async fn compare_delete(
-        &self,
-        pool_guard: &PoolGuard,
-        key: &[Val],
-        row_id: RowID,
-        ignore_del_mask: bool,
-        ts: TrxID,
-    ) -> Result<bool> {
-        debug_assert!(!row_id.is_deleted());
-        if self
-            .mem
-            .lookup_unique(pool_guard, key, row_id, ts)
-            .await?
-            .is_some()
-        {
-            return self
-                .mem
-                .compare_delete(pool_guard, key, row_id, ignore_del_mask, ts)
-                .await;
-        }
-        Ok(true)
-    }
-
-    #[inline]
-    async fn scan_values(
-        &self,
-        pool_guard: &PoolGuard,
-        values: &mut Vec<RowID>,
-        _ts: TrxID,
-    ) -> Result<()> {
-        let mem_entries = self.mem.scan_encoded_entries(pool_guard).await?;
-        let disk_pool_guard = self.disk.disk_pool_guard();
-        let disk = self.disk.open_non_unique(&disk_pool_guard)?;
-        let disk_entries = disk.scan_entries().await?;
-        merge_non_unique_entries(&mem_entries, &disk_entries, values);
-        Ok(())
-    }
-}
-
-/// Composite secondary-index variant for one table secondary index.
-pub(crate) enum DualTreeSecondaryIndex<P: 'static> {
-    /// Composite unique secondary index.
-    Unique(DualTreeUniqueIndex<P>),
-    /// Composite non-unique secondary index.
-    NonUnique(DualTreeNonUniqueIndex<P>),
-}
-
-impl<P: BufferPool> DualTreeSecondaryIndex<P> {
-    /// Pair a freshly built generic MemTree with its cold DiskTree runtime.
-    #[inline]
-    pub(crate) async fn new(
-        mem_index: GenericSecondaryIndex<P>,
-        disk: SecondaryDiskTreeRuntime,
-        pool_guard: &PoolGuard,
-    ) -> Result<Self> {
-        let unique = mem_index.is_unique();
-        if disk.index_no() != mem_index.index_no || disk.is_unique() != unique {
-            let _ = mem_index.destroy(pool_guard).await;
-            return Err(Error::InvalidState);
-        }
-        if unique {
-            let (_, mem) = mem_index.into_unique().ok_or(Error::InvalidState)?;
-            Ok(Self::Unique(DualTreeUniqueIndex { mem, disk }))
-        } else {
-            let (_, mem) = mem_index.into_non_unique().ok_or(Error::InvalidState)?;
-            Ok(Self::NonUnique(DualTreeNonUniqueIndex { mem, disk }))
-        }
-    }
-
-    /// Destroy the mutable MemTree owned by this composite index.
-    #[inline]
-    pub(crate) async fn destroy(self, pool_guard: &PoolGuard) -> Result<()> {
-        match self {
-            Self::Unique(index) => index.destroy(pool_guard).await,
-            Self::NonUnique(index) => index.destroy(pool_guard).await,
-        }
-    }
-
-    /// Return whether this composite index enforces uniqueness.
-    #[inline]
-    pub(crate) fn is_unique(&self) -> bool {
-        matches!(self, Self::Unique(_))
-    }
-
-    /// Return the index number from the cold context.
-    #[inline]
-    pub(crate) fn index_no(&self) -> usize {
-        match self {
-            Self::Unique(index) => index.disk().index_no(),
-            Self::NonUnique(index) => index.disk().index_no(),
-        }
-    }
-
-    /// Returns the unique-index view when this slot is unique.
-    #[inline]
-    pub(crate) fn unique(&self) -> Option<&DualTreeUniqueIndex<P>> {
-        match self {
-            Self::Unique(index) => Some(index),
-            Self::NonUnique(_) => None,
-        }
-    }
-
-    /// Returns the non-unique-index view when this slot is non-unique.
-    #[inline]
-    pub(crate) fn non_unique(&self) -> Option<&DualTreeNonUniqueIndex<P>> {
-        match self {
-            Self::Unique(_) => None,
-            Self::NonUnique(index) => Some(index),
-        }
-    }
-}
-
-#[inline]
-fn merge_unique_entries(
-    mem_entries: &[UniqueMemTreeEntry],
-    disk_entries: &[(Vec<u8>, RowID)],
-    values: &mut Vec<RowID>,
-) {
-    let mut mem_idx = 0;
-    let mut disk_idx = 0;
-    while mem_idx < mem_entries.len() && disk_idx < disk_entries.len() {
-        let mem = &mem_entries[mem_idx];
-        let (disk_key, disk_row_id) = &disk_entries[disk_idx];
-        match mem.encoded_key.as_slice().cmp(disk_key.as_slice()) {
-            Ordering::Less => {
-                values.push(mem.scan_row_id());
-                mem_idx += 1;
-            }
-            Ordering::Equal => {
-                values.push(mem.scan_row_id());
-                mem_idx += 1;
-                disk_idx += 1;
-            }
-            Ordering::Greater => {
-                values.push(*disk_row_id);
-                disk_idx += 1;
-            }
-        }
-    }
-    for mem in &mem_entries[mem_idx..] {
-        values.push(mem.scan_row_id());
-    }
-    for (_, row_id) in &disk_entries[disk_idx..] {
-        values.push(*row_id);
-    }
-}
-
-#[inline]
-fn merge_non_unique_entries(
-    mem_entries: &[NonUniqueMemTreeEntry],
-    disk_entries: &[(Vec<u8>, RowID)],
-    values: &mut Vec<RowID>,
-) {
-    let mut mem_idx = 0;
-    let mut disk_idx = 0;
-    let mut last_emitted_key: Option<Vec<u8>> = None;
-    while mem_idx < mem_entries.len() && disk_idx < disk_entries.len() {
-        let mem = &mem_entries[mem_idx];
-        let (disk_key, disk_row_id) = &disk_entries[disk_idx];
-        match mem.encoded_key.as_slice().cmp(disk_key.as_slice()) {
-            Ordering::Less => {
-                push_active_mem_entry(mem, values, &mut last_emitted_key);
-                mem_idx += 1;
-            }
-            Ordering::Equal => {
-                if !mem.deleted {
-                    push_row_once(&mem.encoded_key, mem.row_id, values, &mut last_emitted_key);
-                }
-                mem_idx += 1;
-                disk_idx += 1;
-            }
-            Ordering::Greater => {
-                push_row_once(disk_key, *disk_row_id, values, &mut last_emitted_key);
-                disk_idx += 1;
-            }
-        }
-    }
-    for mem in &mem_entries[mem_idx..] {
-        push_active_mem_entry(mem, values, &mut last_emitted_key);
-    }
-    for (disk_key, row_id) in &disk_entries[disk_idx..] {
-        push_row_once(disk_key, *row_id, values, &mut last_emitted_key);
-    }
-}
-
-#[inline]
-fn push_active_mem_entry(
-    entry: &NonUniqueMemTreeEntry,
-    values: &mut Vec<RowID>,
-    last_emitted_key: &mut Option<Vec<u8>>,
-) {
-    if !entry.deleted {
-        push_row_once(&entry.encoded_key, entry.row_id, values, last_emitted_key);
-    }
-}
-
-#[inline]
-fn push_row_once(
-    encoded_key: &[u8],
-    row_id: RowID,
-    values: &mut Vec<RowID>,
-    last_emitted_key: &mut Option<Vec<u8>>,
-) {
-    if last_emitted_key.as_deref() != Some(encoded_key) {
-        values.push(row_id);
-        *last_emitted_key = Some(encoded_key.to_vec());
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::buffer::{
-        FixedBufferPool, PoolGuard, PoolRole, global_readonly_pool_scope, table_readonly_pool,
-    };
-    use crate::catalog::{ColumnAttributes, ColumnSpec, IndexAttributes, IndexKey, IndexSpec};
-    use crate::file::build_test_fs;
-    use crate::file::cow_file::SUPER_BLOCK_ID;
-    use crate::file::table_file::{MutableTableFile, TableFile};
-    use crate::index::btree::BTree;
-    use crate::index::btree_key::BTreeKeyEncoder;
-    use crate::index::disk_tree::{NonUniqueDiskTreeExact, UniqueDiskTreePut};
-    use crate::quiescent::QuiescentBox;
-    use crate::value::{ValKind, ValType};
-
-    fn metadata_with_indexes() -> Arc<TableMetadata> {
-        Arc::new(TableMetadata::new(
-            vec![ColumnSpec::new(
-                "c0",
-                ValKind::U32,
-                ColumnAttributes::empty(),
-            )],
-            vec![
-                IndexSpec::new("idx_unique", vec![IndexKey::new(0)], IndexAttributes::UK),
-                IndexSpec::new(
-                    "idx_non_unique",
-                    vec![IndexKey::new(0)],
-                    IndexAttributes::empty(),
-                ),
-            ],
-        ))
-    }
-
-    async fn unique_mem_index(
-        pool: &QuiescentBox<FixedBufferPool>,
-        pool_guard: &PoolGuard,
-    ) -> GenericUniqueBTreeIndex<FixedBufferPool> {
-        let tree = BTree::new(pool.guard(), pool_guard, true, 100)
-            .await
-            .expect("unique MemTree should be created");
-        GenericUniqueBTreeIndex::new(
-            tree,
-            BTreeKeyEncoder::new(vec![ValType::new(ValKind::U32, false)]),
-        )
-    }
-
-    async fn non_unique_mem_index(
-        pool: &QuiescentBox<FixedBufferPool>,
-        pool_guard: &PoolGuard,
-    ) -> GenericNonUniqueBTreeIndex<FixedBufferPool> {
-        let tree = BTree::new(pool.guard(), pool_guard, true, 100)
-            .await
-            .expect("non-unique MemTree should be created");
-        GenericNonUniqueBTreeIndex::new(
-            tree,
-            BTreeKeyEncoder::new(vec![
-                ValType::new(ValKind::U32, false),
-                ValType::new(ValKind::U64, false),
-            ]),
-        )
-    }
-
-    async fn publish_secondary_root(
-        mut mutable: MutableTableFile,
-        index_no: usize,
-        root: BlockID,
-        ts: TrxID,
-    ) -> Arc<TableFile> {
-        mutable
-            .set_secondary_index_root(index_no, root)
-            .expect("test secondary root publication should accept index number");
-        let (table, old_root) = mutable
-            .commit(ts, false)
-            .await
-            .expect("test secondary root publication should commit");
-        drop(old_root);
-        table
-    }
-
-    macro_rules! unique_runtime {
-        ($metadata:ident, $disk_pool:ident) => {
-            UniqueDiskTreeRuntime::new(
-                &$metadata.index_specs[0],
-                $metadata.as_ref(),
-                $disk_pool.file_kind(),
-                Arc::clone($disk_pool.sparse_file()),
-                $disk_pool.global_pool().clone(),
-            )
-            .unwrap()
-        };
-    }
-
-    macro_rules! non_unique_runtime {
-        ($metadata:ident, $disk_pool:ident) => {
-            NonUniqueDiskTreeRuntime::new(
-                &$metadata.index_specs[1],
-                $metadata.as_ref(),
-                $disk_pool.file_kind(),
-                Arc::clone($disk_pool.sparse_file()),
-                $disk_pool.global_pool().clone(),
-            )
-            .unwrap()
-        };
-    }
-
-    #[test]
-    fn test_unique_dual_tree_method_semantics() {
-        smol::block_on(async {
-            let (_temp_dir, fs) = build_test_fs();
-            let metadata = metadata_with_indexes();
-            let table = fs
-                .create_table_file(611, Arc::clone(&metadata), false)
-                .unwrap();
-            let (table, old_root) = table.commit(1, false).await.unwrap();
-            drop(old_root);
-            let global = global_readonly_pool_scope(64 * 1024 * 1024);
-            let disk_pool = table_readonly_pool(&global, 611, &table);
-            let disk_guard = disk_pool.pool_guard();
-            let mut mutable = MutableTableFile::fork(&table, fs.background_writes());
-            let disk_runtime = unique_runtime!(metadata, disk_pool);
-            let disk = disk_runtime.open(SUPER_BLOCK_ID, &disk_guard);
-            let key1 = [Val::from(1u32)];
-            let key2 = [Val::from(2u32)];
-            let key3 = [Val::from(3u32)];
-            let key4 = [Val::from(4u32)];
-            let key5 = [Val::from(5u32)];
-            let key6 = [Val::from(6u32)];
-            let mut writer = disk.batch_writer(&mut mutable, 2);
-            writer
-                .batch_put(&[
-                    UniqueDiskTreePut {
-                        key: &key1,
-                        row_id: 10,
-                    },
-                    UniqueDiskTreePut {
-                        key: &key2,
-                        row_id: 20,
-                    },
-                    UniqueDiskTreePut {
-                        key: &key3,
-                        row_id: 30,
-                    },
-                    UniqueDiskTreePut {
-                        key: &key4,
-                        row_id: 40,
-                    },
-                ])
-                .unwrap();
-            let root = writer.finish().await.unwrap();
-            let table = publish_secondary_root(mutable, 0, root, 2).await;
-
-            let index_pool = QuiescentBox::new(
-                FixedBufferPool::with_capacity(PoolRole::Index, 64 * 1024 * 1024).unwrap(),
-            );
-            let index_guard = (*index_pool).pool_guard();
-            let mem = unique_mem_index(&index_pool, &index_guard).await;
-            assert!(
-                mem.insert_if_not_exists(&index_guard, &key1, 100, false, 3)
-                    .await
-                    .unwrap()
-                    .is_ok()
-            );
-            let runtime = SecondaryDiskTreeRuntime::new(
-                0,
-                Arc::clone(&metadata),
-                Arc::clone(&table),
-                disk_pool.global_pool().clone(),
-            )
-            .unwrap();
-            let index = DualTreeUniqueIndex { mem, disk: runtime };
-
-            assert_eq!(index.disk().published_root().unwrap(), root);
-            assert_eq!(index.disk().index_no(), 0);
-            assert!(
-                index
-                    .mem
-                    .lookup(&index_guard, &key1, 3)
-                    .await
-                    .unwrap()
-                    .is_some()
-            );
-            assert_eq!(
-                index.lookup(&index_guard, &key1, 3).await.unwrap(),
-                Some((100, false))
-            );
-            assert_eq!(
-                index.lookup(&index_guard, &key2, 3).await.unwrap(),
-                Some((20, false))
-            );
-
-            assert!(
-                index
-                    .mask_as_deleted(&index_guard, &key3, 30, 4)
-                    .await
-                    .unwrap()
-            );
-            assert_eq!(
-                index.lookup(&index_guard, &key3, 4).await.unwrap(),
-                Some((30, true))
-            );
-            assert_eq!(
-                index
-                    .insert_if_not_exists(&index_guard, &key4, 400, false, 5)
-                    .await
-                    .unwrap(),
-                IndexInsert::DuplicateKey(40, false)
-            );
-            assert!(
-                index
-                    .insert_if_not_exists(&index_guard, &key5, 50, false, 5)
-                    .await
-                    .unwrap()
-                    .is_ok()
-            );
-            assert_eq!(
-                index
-                    .compare_exchange(&index_guard, &key2, 20, 200, 6)
-                    .await
-                    .unwrap(),
-                IndexCompareExchange::Ok
-            );
-            assert_eq!(
-                index
-                    .compare_exchange(&index_guard, &key4, 999, 9999, 6)
-                    .await
-                    .unwrap(),
-                IndexCompareExchange::Mismatch
-            );
-            assert_eq!(
-                index
-                    .compare_exchange(&index_guard, &key4, 40.deleted(), 400, 6)
-                    .await
-                    .unwrap(),
-                IndexCompareExchange::NotExists
-            );
-            assert_eq!(
-                index
-                    .compare_exchange(&index_guard, &key6, 60, 600, 6)
-                    .await
-                    .unwrap(),
-                IndexCompareExchange::NotExists
-            );
-            assert!(
-                index
-                    .compare_delete(&index_guard, &key4, 40, false, 7)
-                    .await
-                    .unwrap()
-            );
-            assert!(
-                index
-                    .compare_delete(&index_guard, &key4, 41, false, 7)
-                    .await
-                    .unwrap()
-            );
-
-            let mut values = Vec::new();
-            index
-                .scan_values(&index_guard, &mut values, 8)
-                .await
-                .unwrap();
-            assert_eq!(values, vec![100, 200, 30u64.deleted(), 40, 50]);
-
-            let unchanged_disk = disk_runtime.open(root, &disk_guard);
-            assert_eq!(unchanged_disk.lookup(&key2).await.unwrap(), Some(20));
-            assert_eq!(unchanged_disk.lookup(&key3).await.unwrap(), Some(30));
-            assert_eq!(unchanged_disk.lookup(&key4).await.unwrap(), Some(40));
-        });
-    }
-
-    #[test]
-    fn test_disk_runtime_resolves_published_root_per_open() {
-        smol::block_on(async {
-            let (_temp_dir, fs) = build_test_fs();
-            let metadata = metadata_with_indexes();
-            let table = fs
-                .create_table_file(614, Arc::clone(&metadata), false)
-                .unwrap();
-            let (table, old_root) = table.commit(1, false).await.unwrap();
-            drop(old_root);
-            let global = global_readonly_pool_scope(64 * 1024 * 1024);
-            let disk_pool = table_readonly_pool(&global, 614, &table);
-            let disk_guard = disk_pool.pool_guard();
-            let key1 = [Val::from(1u32)];
-            let key2 = [Val::from(2u32)];
-
-            let mut mutable = MutableTableFile::fork(&table, fs.background_writes());
-            let disk_runtime = unique_runtime!(metadata, disk_pool);
-            let disk = disk_runtime.open(SUPER_BLOCK_ID, &disk_guard);
-            let root_a = {
-                let mut writer = disk.batch_writer(&mut mutable, 2);
-                writer
-                    .batch_put(&[UniqueDiskTreePut {
-                        key: &key1,
-                        row_id: 10,
-                    }])
-                    .unwrap();
-                writer.finish().await.unwrap()
-            };
-            let table = publish_secondary_root(mutable, 0, root_a, 2).await;
-            let runtime = SecondaryDiskTreeRuntime::new(
-                0,
-                Arc::clone(&metadata),
-                Arc::clone(&table),
-                disk_pool.global_pool().clone(),
-            )
-            .unwrap();
-            assert_eq!(runtime.published_root().unwrap(), root_a);
-
-            let opened_a_guard = runtime.disk_pool_guard();
-            let opened_a = runtime.open_unique(&opened_a_guard).unwrap();
-            assert_eq!(opened_a.lookup(&key1).await.unwrap(), Some(10));
-            assert_eq!(opened_a.lookup(&key2).await.unwrap(), None);
-
-            let mut mutable = MutableTableFile::fork(&table, fs.background_writes());
-            let disk = disk_runtime.open(root_a, &disk_guard);
-            let root_b = {
-                let mut writer = disk.batch_writer(&mut mutable, 3);
-                writer
-                    .batch_put(&[UniqueDiskTreePut {
-                        key: &key2,
-                        row_id: 20,
-                    }])
-                    .unwrap();
-                writer.finish().await.unwrap()
-            };
-            assert_ne!(root_a, root_b);
-            let table_after_b = publish_secondary_root(mutable, 0, root_b, 3).await;
-            assert_eq!(table_after_b.active_root().secondary_index_roots[0], root_b);
-            assert_eq!(runtime.published_root().unwrap(), root_b);
-
-            assert_eq!(opened_a.lookup(&key1).await.unwrap(), Some(10));
-            assert_eq!(opened_a.lookup(&key2).await.unwrap(), None);
-
-            let opened_b_guard = runtime.disk_pool_guard();
-            let opened_b = runtime.open_unique(&opened_b_guard).unwrap();
-            assert_eq!(opened_b.lookup(&key1).await.unwrap(), Some(10));
-            assert_eq!(opened_b.lookup(&key2).await.unwrap(), Some(20));
-        });
-    }
-
-    #[test]
-    fn test_non_unique_dual_tree_merge_and_overlay_semantics() {
-        smol::block_on(async {
-            let (_temp_dir, fs) = build_test_fs();
-            let metadata = metadata_with_indexes();
-            let table = fs
-                .create_table_file(612, Arc::clone(&metadata), false)
-                .unwrap();
-            let (table, old_root) = table.commit(1, false).await.unwrap();
-            drop(old_root);
-            let global = global_readonly_pool_scope(64 * 1024 * 1024);
-            let disk_pool = table_readonly_pool(&global, 612, &table);
-            let disk_guard = disk_pool.pool_guard();
-            let mut mutable = MutableTableFile::fork(&table, fs.background_writes());
-            let disk_runtime = non_unique_runtime!(metadata, disk_pool);
-            let disk = disk_runtime.open(SUPER_BLOCK_ID, &disk_guard);
-            let key1 = [Val::from(1u32)];
-            let key2 = [Val::from(2u32)];
-            let key3 = [Val::from(3u32)];
-            let mut writer = disk.batch_writer(&mut mutable, 2);
-            writer
-                .batch_insert(&[
-                    NonUniqueDiskTreeExact {
-                        key: &key1,
-                        row_id: 10,
-                    },
-                    NonUniqueDiskTreeExact {
-                        key: &key1,
-                        row_id: 11,
-                    },
-                    NonUniqueDiskTreeExact {
-                        key: &key2,
-                        row_id: 20,
-                    },
-                    NonUniqueDiskTreeExact {
-                        key: &key3,
-                        row_id: 30,
-                    },
-                ])
-                .unwrap();
-            let root = writer.finish().await.unwrap();
-            let table = publish_secondary_root(mutable, 1, root, 2).await;
-
-            let index_pool = QuiescentBox::new(
-                FixedBufferPool::with_capacity(PoolRole::Index, 64 * 1024 * 1024).unwrap(),
-            );
-            let index_guard = (*index_pool).pool_guard();
-            let mem = non_unique_mem_index(&index_pool, &index_guard).await;
-            assert!(
-                mem.insert_if_not_exists(&index_guard, &key1, 12, false, 3)
-                    .await
-                    .unwrap()
-                    .is_ok()
-            );
-            let runtime = SecondaryDiskTreeRuntime::new(
-                1,
-                Arc::clone(&metadata),
-                Arc::clone(&table),
-                disk_pool.global_pool().clone(),
-            )
-            .unwrap();
-            let index = DualTreeNonUniqueIndex { mem, disk: runtime };
-
-            assert_eq!(index.disk().published_root().unwrap(), root);
-            assert_eq!(index.disk().index_no(), 1);
-            assert!(
-                index
-                    .mem
-                    .lookup_unique(&index_guard, &key1, 12, 3)
-                    .await
-                    .unwrap()
-                    .is_some()
-            );
-            assert!(
-                index
-                    .mask_as_deleted(&index_guard, &key1, 10, 4)
-                    .await
-                    .unwrap()
-            );
-            let mut rows = Vec::new();
-            index
-                .lookup(&index_guard, &key1, &mut rows, 4)
-                .await
-                .unwrap();
-            assert_eq!(rows, vec![11, 12]);
-            assert_eq!(
-                index
-                    .lookup_unique(&index_guard, &key1, 10, 4)
-                    .await
-                    .unwrap(),
-                Some(false)
-            );
-            assert_eq!(
-                index
-                    .lookup_unique(&index_guard, &key1, 11, 4)
-                    .await
-                    .unwrap(),
-                Some(true)
-            );
-            assert_eq!(
-                index
-                    .lookup_unique(&index_guard, &key1, 99, 4)
-                    .await
-                    .unwrap(),
-                None
-            );
-            assert_eq!(
-                index
-                    .insert_if_not_exists(&index_guard, &key1, 11, false, 5)
-                    .await
-                    .unwrap(),
-                IndexInsert::DuplicateKey(11, false)
-            );
-            assert!(
-                index
-                    .mask_as_active(&index_guard, &key1, 10, 5)
-                    .await
-                    .unwrap()
-            );
-            rows.clear();
-            index
-                .lookup(&index_guard, &key1, &mut rows, 5)
-                .await
-                .unwrap();
-            assert_eq!(rows, vec![10, 11, 12]);
-            assert!(
-                index
-                    .insert_if_not_exists(&index_guard, &key1, 13, false, 6)
-                    .await
-                    .unwrap()
-                    .is_ok()
-            );
-            assert!(
-                index
-                    .mask_as_deleted(&index_guard, &key2, 20, 7)
-                    .await
-                    .unwrap()
-            );
-            assert!(
-                index
-                    .compare_delete(&index_guard, &key3, 30, false, 7)
-                    .await
-                    .unwrap()
-            );
-
-            let mut values = Vec::new();
-            index
-                .scan_values(&index_guard, &mut values, 8)
-                .await
-                .unwrap();
-            assert_eq!(values, vec![10, 11, 12, 13, 30]);
-
-            let unchanged_disk = disk_runtime.open(root, &disk_guard);
-            assert_eq!(
-                unchanged_disk.prefix_scan(&key1).await.unwrap(),
-                vec![10, 11]
-            );
-            assert_eq!(unchanged_disk.prefix_scan(&key2).await.unwrap(), vec![20]);
-        });
-    }
-
-    #[test]
-    fn test_dual_tree_secondary_index_wrapper() {
-        smol::block_on(async {
-            let (_temp_dir, fs) = build_test_fs();
-            let metadata = metadata_with_indexes();
-            let table = fs
-                .create_table_file(613, Arc::clone(&metadata), false)
-                .unwrap();
-            let (table, old_root) = table.commit(1, false).await.unwrap();
-            drop(old_root);
-            let global = global_readonly_pool_scope(64 * 1024 * 1024);
-            let disk_pool = table_readonly_pool(&global, 613, &table);
-            let index_pool = QuiescentBox::new(
-                FixedBufferPool::with_capacity(PoolRole::Index, 64 * 1024 * 1024).unwrap(),
-            );
-            let index_guard = (*index_pool).pool_guard();
-            let mem = unique_mem_index(&index_pool, &index_guard).await;
-            let shadow_key = [Val::from(9u32)];
-            assert!(
-                mem.insert_delete_shadow_if_absent(&index_guard, &shadow_key, 90, 2)
-                    .await
-                    .unwrap()
-            );
-            let runtime = SecondaryDiskTreeRuntime::new(
-                0,
-                Arc::clone(&metadata),
-                Arc::clone(&table),
-                disk_pool.global_pool().clone(),
-            )
-            .unwrap();
-            let composite =
-                DualTreeSecondaryIndex::Unique(DualTreeUniqueIndex { mem, disk: runtime });
-            assert!(composite.is_unique());
-            assert_eq!(composite.index_no(), 0);
-
-            let mem = non_unique_mem_index(&index_pool, &index_guard).await;
-            let runtime = SecondaryDiskTreeRuntime::new(
-                1,
-                Arc::clone(&metadata),
-                Arc::clone(&table),
-                disk_pool.global_pool().clone(),
-            )
-            .unwrap();
-            let composite =
-                DualTreeSecondaryIndex::NonUnique(DualTreeNonUniqueIndex { mem, disk: runtime });
-            assert!(!composite.is_unique());
-            assert_eq!(composite.index_no(), 1);
-        });
-    }
-}
